Extract AP mode configuration out of public API
This also breaks the function up into a pure part and one with side
effects to support testing.
